+++ src/features/resource-calendar/components/day-view/resource-day-horizontal.tsx
@@ -12,7 +12,7 @@ export const ResourceDayHorizontal: React.FC = () => {
     stickyViewHeader,
     viewHeaderClassName,
     currentLocale,
-    is24Hour,
+    timeFormat,
   } = useResourceCalendarContext()
 
   // Generate time columns (hourly slots)
@@ -63,8 +63,7 @@ export const ResourceDayHorizontal: React.FC = () => {
                     >
                       {Intl.DateTimeFormat(currentLocale, {
                         hour: 'numeric',
-                        minute: 'numeric',
-                        hour12: !is24Hour,
+                        hour12: timeFormat === '12-hour',
                       }).format(col.toDate())}
                     </motion.div>
                   </AnimatePresence>